Move functions for executing processes into separate adapter NOTE: This is a temporary change that can be amended, once the consumer api can authenticate itself against the external authority. Until then, we need this adapter in order to be able to get a full process model, regardless of the users current permissions. That is, because the consumer api now uses the executeProcessService, which requires a full process model to run properly.
